project/requirements.txt
fastapi
uvicorn
pydantic
motor
openai>=1.0.0
google-generativeai
httpx
python-dotenv
